Do not cache complex generators, except in the TestSuite plugin
So this patch ensures that:

a) --no-cache completely bypasses the cachedecorator code.
b) complex generators (generators with subgenerators) are not cached,
in case the subgenerators have been traversed/exhausted by the time they
come to be dumped.
c) Allow caching of complex generators by the TestSuite plugin when we
know it won't use up any subgenerators.

The decorator wrapper now checks config.NO_CACHE and passes it off to
the actual wrapper.  This must be done in the wrapper rather than in the
decorator, because the decorator is run at import time, so changes to
the NO_CACHE value will not be seen later on.

The only hackish bit of this patch is setting the NO_CACHE value.  If it
isn't set just before the wrapper is called, the value seems to get
lost somewhere along the way.