Move a couple tasks to Longterm.
I think the win 8/2012 support is a bit of a big ask for 2.2.  I'd save that for a 2.3/3.0 one.  Similarly for application profiles.  2.2 is mostly supposed to be for getting linux/mac support out without holding up 2.1.  The hope is for a release in a very short time (3 months, realistically slipping to 6 months).  The others should probably be saved for 2.3 (although by all means worked upon/designed starting now).